Orders Controller Implementation

+++ RestaurantAPI.Core.Application/Dtos/Dish/DishDto.cs
@@ -1,5 +1,4 @@
 ﻿using RestaurantAPI.Core.Application.Dtos.Ingredient;
-using RestaurantAPI.Core.Application.Enums;
 
 namespace RestaurantAPI.Core.Application.Dtos.Dish
 {
@@ -7,7 +6,7 @@ namespace RestaurantAPI.Core.Application.Dtos.Dish
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public decimal Price { get; set; }
+        public double Price { get; set; }
         public int ServesPeople { get; set; }
         public List<IngredientDto> Ingredients { get; set; }
         public string Category { get; set; }
